mqtt_aggregator/requirements.txt
cherrypy
paho_mqtt==1.6.1
requests